[receiver/datadog] Fix import cycle
