+++ cmd/otelcontribcol/Dockerfile
@@ -7,7 +7,7 @@ ARG USER_UID=10001
 USER ${USER_UID}
 
 COPY --from=prep /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/ca-certificates.crt
-COPY otelcontribcol /
+COPY otelcol-contrib /
 EXPOSE 4317 55680 55679
-ENTRYPOINT ["/otelcontribcol"]
+ENTRYPOINT ["/otelcol-contrib"]
 CMD ["--config", "/etc/otel/config.yaml"]
